Separate step one tests into individual tests [x] separated all the substeps into separate tests [x] moved one css selector from OCC file to common file
